% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chisqSimShiny}
\alias{chisqSimShiny}
\title{Chi-Square Simulation (Contingency Table)}
\usage{
chisqSimShiny(x, data = parent.frame())
}
\arguments{
\item{x}{Could be a formula.  If so, it should be of the form ~var1+var2.
Otherwise either a table or matrix of summary data.}

\item{data}{dataframe supplying variables for formula x.  If variables in x are not found in the data,
then they will be searched for in the parent environment.}
}
\value{
side effects
}
\description{
Perform chi-square test for association, by simulation.  Enter either
formula-data input or a summary table.
}
\note{
This is a locally-run Shiny app.  It may not work properly on some R Studio Server set-ups,
especially on the CentoOS operating system.
}
\examples{
\dontrun{
# from a data frame:
chisqSimShiny(~am+cyl,data=mtcars)

# from a summary table:
DoesNotSmoke <- c(NeitherSmokes=1168,OneSmokes=1823,BothSmoke=1380)
Smokes <- c(188,416,400)
ChildParents <- rbind(DoesNotSmoke,Smokes)
chisqSimShiny(ChildParents)
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

