% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cloudSh}
\alias{cloudSh}
\title{3-D Scatter Plots}
\usage{
cloudSh(x, data = parent.frame(),group=NULL)
}
\arguments{
\item{x}{Could be a formula.  If so, it should be of the form z~x*y|g1*g2, as in the
\code{lattice} function \code{cloud}.}

\item{data}{data frame supplying variables for formula x.  If variables in x are not found in the data,
then they will be searched for in the parent environment.}

\item{group}{A factor variable by which to groups the points of the plot.}
}
\value{
side effects
}
\description{
Uses the \code{cloud} function from the \code{lattice} package.
}
\note{
This is a locally-run Shiny app.  It may not work properly on some R Studio Server set-ups,
especially on the CentOS operating system.  For best views, open the app in the browser.
}
\examples{
\dontrun{
cloudSh(Sepal.Length~Petal.Length*Petal.Width,data=iris)
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

