% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatterDF}
\alias{scatterDF}
\title{Dynamic Faceting in Scatter Plots}
\usage{
scatterDF(form, data = parent.frame())
}
\arguments{
\item{form}{a formula, of the form y~x|z, where z is the facetting variable.}

\item{data}{dataframe supplying variables for formula.  If variables in the formula are not found in the data,
then they will be searched for in the parent environment.}
}
\value{
side effects
}
\description{
Interactively facet a scatterplot by a third variable.
}
\note{
This is a locally-run Shiny app.  It may not work properly on some R Studio Server set-ups,
especially on the CentOS operating system.
}
\examples{
\dontrun{
scatterDF(Petal.Length ~ Petal.Width | Species, data = iris)
if (require(mosaicData)) scatterDF(sat ~ salary | frac, data = SAT)
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

