% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test}
\alias{test}
\title{Test}
\usage{
test()
}
\description{
To show difference in screen width when function is loaded and called vs.
 called from an attached package.
}
\examples{
\dontrun{
test()
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

